# Cache classes editor help (a.k.a. faster editor startup)