--- a/editor/editor_help.h
+++ b/editor/editor_help.h
@@ -193,7 +193,9 @@ class EditorHelp : public VBoxContainer {
 	static Thread thread;
 
 	static void _wait_for_thread();
+	static void _load_doc_thread(void *p_udata);
 	static void _gen_doc_thread(void *p_udata);
+	static void _generate_doc_first_step();
 
 protected:
 	virtual void _update_theme_item_cache() override;
@@ -205,6 +207,7 @@ public:
 	static void generate_doc();
 	static DocTools *get_doc_data();
 	static void cleanup_doc();
+	static String get_cache_full_path();
 
 	void go_to_help(const String &p_help);
 	void go_to_class(const String &p_class, int p_scroll = 0);