(Built on top of #71644, which is included here. Please review newest commit only.)

The class reference is now cached and regenerated every time the doc tool is run or the commit hash changes. This avoids, on every regular run of the editor, paying the cost of instantiating all the known classes at startup.

Docs generation have two steps: a synchronous one, which just contributes to the perceived startup time, and one which runs on a thread, whose cost is negligible, but I guess can affect some low performance systems for the brief time it lasts.

On my system (AMD Ryzen 9 5900HX with NVMe SSD), a regular startup gives these times:
- Before this PR: 240 ms sync + 114 ms async.
- With this PR: 0 ms sync + 200 ms async.

In short, on my system the Godot editor takes 0.25 seconds shorter to start. Not a lot, but good in any case.