@akien-mga 2023-04-13
Did some before/after testing, with a naive approach timing the editor startup time on a simple/near empty project.

System:
```
$ inxi -C
CPU:
  Info: quad core model: Intel Core i7-8705G bits: 64 type: MT MCP cache:
    L2: 1024 KiB
  Speed (MHz): avg: 1726 min/max: 800/4100 cores: 1: 900 2: 913 3: 900
    4: 3100 5: 900 6: 900 7: 3100 8: 3100
```
Godot build: `scons target=editor`

For each test I removed `~/.cache/godot`, `~/.local/share/godot/app_userdata/<project name>` and `<project path>/.godot`.
And then I timed the startup of the editor when using the `--quit` argument, so it quits as soon as it's initialized:
```
time ./bin/godot.linuxbsd.editor.x86_64 -e --path ~/New\ Game\ Project4.0/ --quit
```

### Before this PR (68b8156fe3075e0e45b070d26fce2a0baae6f293)

Cold run, generates shader cache, etc. (average of 3 runs after 1 warmup run):
8.765s

Subsequent runs (average of 20, after 3 warmup runs):
```
$ hyperfine -i -w3 -m20 "./bin/godot.linuxbsd.editor.x86_64.old -e --path ~/New\ Game\ Project4.0/ --quit"
Benchmark 1: ./bin/godot.linuxbsd.editor.x86_64.old -e --path ~/New\ Game\ Project4.0/ --quit
  Time (mean ± σ):      6.623 s ±  0.264 s    [User: 5.017 s, System: 0.469 s]
  Range (min … max):    5.559 s …  6.819 s    20 runs
```

### With this PR (rebased on 68b8156f)

Cold run, generates shader cache, etc. + doc cache (average of 3 runs after 1 warmup run):
8.655s

Subsequent runs (average of 20, after 3 warmup runs):
```
hyperfine -i -w3 -m20 "./bin/godot.linuxbsd.editor.x86_64 -e --path ~/New\ Game\ Project4.0/ --quit"
Benchmark 1: ./bin/godot.linuxbsd.editor.x86_64 -e --path ~/New\ Game\ Project4.0/ --quit
  Time (mean ± σ):      6.282 s ±  0.238 s    [User: 4.806 s, System: 0.485 s]
  Range (min … max):    5.332 s …  6.459 s    20 runs
```

So it does seem to shave off around 350 ms in average for me, though it's not a very reliable test. Interesting that the cold run also seems to be tangentially faster / same duration, so saving the cache doesn't seem to be slow.