@RedworkDE 2023-04-19
If an outdated cache exists, the `class_list` is missing a few entries. Notably `@GlobalScope` being missing causes #76250.

```diff
diff --git a/class_list b/class_list
index fe8c67b9c65..2ec0a0a7c4e 100644
--- a/diff-tmp
+++ b/diff-tmp
@@ -1,5 +1,3 @@
-@GDScript
-@GlobalScope
 AABB
 AcceptDialog
 AESContext
@@ -54,7 +52,6 @@ Area2D
 Area2D.SpaceOverride
 Area3D
 Area3D.SpaceOverride
-Array
 ArrayMesh
 ArrayOccluder3D
 AspectRatioContainer
@@ -292,7 +289,6 @@ CylinderShape3D
 DampedSpringJoint2D
 Decal
 Decal.DecalTexture
-Dictionary
 DirAccess
 DirectionalLight2D
 DirectionalLight3D
@@ -402,7 +398,6 @@ FileDialog
 FileDialog.AccessEnum
 FileDialog.FileModeEnum
 FileSystemDock
-float
 FlowContainer
 FlowContainer.AlignmentMode
 FogMaterial
@@ -668,7 +663,6 @@ Node2D
 Node3D
 Node3D.RotationEditModeEnum
 Node3DGizmo
-NodePath
 Noise
 NoiseTexture2D
 Object
@@ -697,19 +691,10 @@ ORMMaterial3D
 OS
 OS.RenderingDriver
 OS.SystemDir
-PackedByteArray
-PackedColorArray
 PackedDataContainer
 PackedDataContainerRef
-PackedFloat32Array
-PackedFloat64Array
-PackedInt32Array
-PackedInt64Array
 PackedScene
 PackedScene.GenEditState
-PackedStringArray
-PackedVector2Array
-PackedVector3Array
 PacketPeer
 PacketPeerDTLS
 PacketPeerDtls.Status
@@ -771,7 +756,6 @@ PhysicsServer2D.PinJointParam
 PhysicsServer2D.ProcessInfo
 PhysicsServer2D.ShapeType
 PhysicsServer2D.SpaceParameter
-PhysicsServer2DExtension
 PhysicsServer2DManager
 PhysicsServer3D
 PhysicsServer3D.AreaBodyStatus
@@ -793,7 +777,6 @@ PhysicsServer3D.ProcessInfo
 PhysicsServer3D.ShapeType
 PhysicsServer3D.SliderJointParam
 PhysicsServer3D.SpaceParameter
-PhysicsServer3DExtension
 PhysicsServer3DManager
 PhysicsServer3DRenderingServerHandler
 PhysicsShapeQueryParameters2D
@@ -996,7 +979,6 @@ RichTextEffect
 RichTextLabel
 RichTextLabel.ListType
 RichTextLabel.MenuItems
-RID
 RigidBody2D
 RigidBody2D.CcdMode
 RigidBody2D.CenterOfMassModeEnum
@@ -1044,7 +1026,6 @@ ShapeCast2D
 ShapeCast3D
 short
 Shortcut
-Signal
 Skeleton2D
 Skeleton3D
 SkeletonIK3D
@@ -1095,8 +1076,6 @@ StreamPeerTCP
 StreamPeerTcp.Status
 StreamPeerTLS
 StreamPeerTls.Status
-String
-StringName
 StyleBox
 StyleBoxEmpty
 StyleBoxFlat
@@ -1217,7 +1196,6 @@ Upnp.UpnpResult
 UPNPDevice
 UpnpDevice.IgdStatusEnum
 ushort
-Variant
 VBoxContainer
 Vector2
 Vector2i
```

---

Also random thought, but I don't think switch between different build configurations (ie mono and standard version) of the same commit or having uncommitted changes invalidates the cache.